pyproject.toml
[project]
name = "socrates"
version = "0.2.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "maxapi>=0.9.15",
    "orjson>=3.11.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "uv"
update_changelog_on_bump = false

[tool.ruff]
builtins = ["_"]
fix = true
src = ["bot"]
target-version = "py314"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.ruff.lint.isort]
lines-after-imports = 2
lines-between-types = 1